Extend recorder component retention
This is a pre-requisite for having home assistant itself make the graphs
for temperature tracking.  Which itself is a pre-requisite for removing
the smartthings-mqtt-bridge component and instead use the HASS
smartthings native integration.  Which itself seems like a good idea,
given the relative fragility of the pm2 daemon at startup.