Re-enable new authentication
By adding only the legacy_api login mechanism, I lost the ability to
login with the homeassistant configured users.
So I've added in the configuration to re-enable the homeassistant auth
mechanism, overriding the legacy one (which is deprecated and will be
removed in future - yay, I'm ahead of the curve on this one!).