Exclude input_boolean.kitmocorr from config
It's not a config input boolean; it's only there to shadow the correct
state of the Smartthing's motion detector which constantly signals that
motion is occurring for at least 5 seconds after motion has stopped.
This is fatal, as the controlled lights flash on and off whilst motion
is being reported.